(Positional embedding) Cast pe_length to jnp.int32
pe_length was usually a jnp.uint32 which would've caused incorrect
behaviour when we negated it in the jnp.roll() call.

Also prevents a cryptic XLA error about attempting to add a u32 and an
s32 (uint32 and int32), so I guess it had to be a signed integer
regardless.